# optimization: Optimize IsSpace function for common non-whitespace characters