@paplorinc 2024-03-15
> Is this used in a hot path that is relevant, to warrant the changes?

It's used in [base 58 and 16 conversions](https://github.com/search?q=repo%3Abitcoin%2Fbitcoin+%2F%28%3F-i%29IsSpace%2F+language%3AC%2B%2B&type=code) while iterating over each character of the input.
We're expecting most of them to not be spaces, it would make sense to optimize for that scenario and exit early.