@paplorinc 2024-03-15
It's a small improvement, indeed, but it eliminates useless work, done frequently.
The exhaustive tests make sure the behavior is exactly the same as before, so while it may be considered low-reward, it should be low-risk as well.
Is thereany area (testing, documentation, code) that you would like me to  improve to increase the risk/reward ratio?