@laanwj 2024-04-09
Is this really a bottleneck? A benchmark would be useful to see if this isn't something the compiler already does.

ACK on the test.