@paplorinc 2024-04-09
Thanks for the review @laanwj. The gain is small, but measurable:
> make && ./src/bench/bench_bitcoin --filter=HexParse --min-time=10000


`Before:`
```
|           ns/base16 |            base16/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.68 |    1,461,990,359.30 |    0.0% |     10.83 | `HexParse`
|                0.68 |    1,461,442,745.57 |    0.1% |     10.83 | `HexParse`
|                0.68 |    1,461,262,276.18 |    0.1% |     10.83 | `HexParse`
```

`After:`
```
|           ns/base16 |            base16/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                0.67 |    1,495,985,893.94 |    0.0% |     10.82 | `HexParse`
|                0.67 |    1,494,916,055.36 |    0.1% |     10.82 | `HexParse`
|                0.67 |    1,494,545,662.66 |    0.0% |     10.82 | `HexParse`
```
which results in a speedup of ~2% for 130 non-whitespace hexadecimal characters:
```
(1,495,985,893.94 / 1,461,990,359.30) = 2.3%
```